Improve File System of Employee and restructure of views
-Move file folder to static folder
-Separate "detail" to multiple views
-Change model of "File"
-Add placeholder to field of form views

+++ setup.py
@@ -1,28 +1,30 @@
 # coding:utf-8
 from citywok_manager import create_app, db, bcrypt
-from citywok_manager.models import Country, User, Job, Employee
+from citywok_manager.models import Country, User, Job
 import os
 import shutil
-
-app = create_app()
 
 
 def clean_path(path_to_clean):
     for filename in os.listdir(path_to_clean):
         file_path = os.path.join(path_to_clean, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print('Failed to delete %s. Reason: %s' % (file_path, e))
+        if filename != '.gitkeep':
+            try:
+                if os.path.isfile(file_path) or os.path.islink(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
+            except Exception as e:
+                print('Failed to delete %s. Reason: %s' % (file_path, e))
+
+
+app = create_app()
 
 
 with app.app_context():
 
-    clean_path(os.path.join(app.root_path, 'download', 'employee'))
-    clean_path(os.path.join(app.root_path, 'download', 'supplier'))
+    clean_path(os.path.join(app.config['EMPLOYEE_FILE']))
+    clean_path(os.path.join(app.config['SUPPLIER_FILE']))
 
     # clear the database if needed
     db.drop_all()
